add cron service on flag in

+++ src/utills/task-service.service.ts
@@ -9,8 +9,10 @@ export class TasksService {
 
   @Cron(CronExpression.EVERY_30_SECONDS)
   async handleCron() {
-    this.logger.debug('Operation start');
-    await this.bookingRepository.removeUnusedBooking();
-    this.logger.debug('Operation completed');
+    if (process.env.CRON_JOB_STATUS == 'ON') {
+      this.logger.debug('Operation start');
+      await this.bookingRepository.removeUnusedBooking();
+      this.logger.debug('Operation completed');
+    }
   }
 }
